requirements.txt
Django
django-filter
django-dbbackup
django-allauth